fix batch module project references
