+++ source/dotnet/wholesale-api/Batches/Batches.Application/BatchAggregate/Batch.cs
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Batches.Infrastructure.GridAreaAggregate;
+using Energinet.DataHub.Wholesale.Batches.Application.GridAreaAggregate;
 using Energinet.DataHub.Wholesale.Batches.Interfaces;
 using Energinet.DataHub.Wholesale.Batches.Interfaces.Models;
 using NodaTime;
 
-namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
+namespace Energinet.DataHub.Wholesale.Batches.Application.BatchAggregate;
 
 public class Batch
 {
