+++ source/dotnet/wholesale-api/Batches/Batches.Application/BatchAggregate/BatchExecutionState.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
+namespace Energinet.DataHub.Wholesale.Batches.Application.BatchAggregate;
 
 /// <summary>
 /// IMPORTANT: Do not change numeric values as it'll affect persistence or communication.
