+++ source/dotnet/wholesale-api/Batches/Batches.Application/Batches.Application.csproj
@@ -12,7 +12,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Batches.Interfaces\Batches.Interfaces.csproj" />
-    <ProjectReference Include="..\Batches.Infrastructure\Batches.Infrastructure.csproj" />
   </ItemGroup>
 
 </Project>
