+++ source/dotnet/wholesale-api/Batches/Batches.Application/CalculationState.cs
@@ -12,6 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
+namespace Energinet.DataHub.Wholesale.Batches.Application;
 
-public sealed record CalculationId(long Id);
+public enum CalculationState
+{
+    Pending,
+    Running,
+    Completed,
+    Canceled,
+    Failed,
+}
