+++ source/dotnet/wholesale-api/Batches/Batches.Application/CreateBatchHandler.cs
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
-using Energinet.DataHub.Wholesale.Batches.Infrastructure.Persistence;
+using Energinet.DataHub.Wholesale.Batches.Application.BatchAggregate;
 using Energinet.DataHub.Wholesale.Batches.Interfaces;
 
 namespace Energinet.DataHub.Wholesale.Batches.Application;
