+++ source/dotnet/wholesale-api/Batches/Batches.Application/GridAreaAggregate/GridAreaCode.cs
@@ -15,7 +15,7 @@
 using System.Text.RegularExpressions;
 using Energinet.DataHub.Wholesale.Batches.Interfaces;
 
-namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.GridAreaAggregate;
+namespace Energinet.DataHub.Wholesale.Batches.Application.GridAreaAggregate;
 
 public sealed record GridAreaCode
 {
