+++ source/dotnet/wholesale-api/Batches/Batches.Application/IBatchExecutionStateDomainService.cs
@@ -10,15 +10,11 @@
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
-// limitations under the License.
+// limitations under the License.using Energinet.DataHub.Wholesale.Common.JobRunner;
 
-namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.CalculationDomainService;
+namespace Energinet.DataHub.Wholesale.Batches.Application;
 
-public enum CalculationState
+public interface IBatchExecutionStateDomainService
 {
-    Pending,
-    Running,
-    Completed,
-    Canceled,
-    Failed,
+    Task UpdateExecutionStateAsync();
 }
