+++ source/dotnet/wholesale-api/Batches/Batches.Application/IBatchRepository.cs
@@ -12,10 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Batches.Infrastructure.GridAreaAggregate;
+using Energinet.DataHub.Wholesale.Batches.Application.BatchAggregate;
+using Energinet.DataHub.Wholesale.Batches.Application.GridAreaAggregate;
 using NodaTime;
 
-namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
+namespace Energinet.DataHub.Wholesale.Batches.Application;
 
 public interface IBatchRepository
 {
