+++ source/dotnet/wholesale-api/Batches/Batches.Application/ICalculationDomainService.cs
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
+using Energinet.DataHub.Wholesale.Batches.Application.BatchAggregate;
 
-namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.CalculationDomainService;
+namespace Energinet.DataHub.Wholesale.Batches.Application;
 
 public interface ICalculationDomainService
 {
