+++ source/dotnet/wholesale-api/Batches/Batches.Application/IUnitOfWork.cs
@@ -10,11 +10,11 @@
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
-// limitations under the License.using Energinet.DataHub.Wholesale.Common.JobRunner;
+// limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchExecutionStateDomainService;
+namespace Energinet.DataHub.Wholesale.Batches.Application;
 
-public interface IBatchExecutionStateDomainService
+public interface IUnitOfWork
 {
-    Task UpdateExecutionStateAsync();
+    Task CommitAsync();
 }
