+++ source/dotnet/wholesale-api/Batches/Batches.Infrastructure/CalculationDomainService/ICalculationEngineClient.cs
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
+using Energinet.DataHub.Wholesale.Batches.Application;
+using Energinet.DataHub.Wholesale.Batches.Application.BatchAggregate;
 
 namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.CalculationDomainService;
