+++ source/dotnet/wholesale-api/Batches/Batches.Infrastructure/Persistence/DatabaseContext.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using System.Diagnostics.CodeAnalysis;
-using Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
+using Energinet.DataHub.Wholesale.Batches.Application.BatchAggregate;
 using Energinet.DataHub.Wholesale.Batches.Infrastructure.Persistence.Batches;
 using Microsoft.EntityFrameworkCore;
