+++ source/dotnet/wholesale-api/Batches/Batches.Infrastructure/Persistence/IDatabaseContext.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
+using Energinet.DataHub.Wholesale.Batches.Application.BatchAggregate;
 using Microsoft.EntityFrameworkCore;
 
 namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.Persistence;
