+++ source/dotnet/wholesale-api/Batches/Batches.IntegrationTests/Infrastructure/Persistence/Batch/BatchRepositoryTests.cs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
-using Energinet.DataHub.Wholesale.Batches.Infrastructure.GridAreaAggregate;
+using Energinet.DataHub.Wholesale.Batches.Application.BatchAggregate;
+using Energinet.DataHub.Wholesale.Batches.Application.GridAreaAggregate;
 using Energinet.DataHub.Wholesale.Batches.Infrastructure.Persistence.Batches;
 using Energinet.DataHub.Wholesale.Batches.IntegrationTests.Fixture.Database;
 using Energinet.DataHub.Wholesale.Batches.Interfaces.Models;
@@ -198,7 +198,7 @@ public class BatchRepositoryTests : IClassFixture<WholesaleDatabaseFixture>
         await using var writeContext = _databaseManager.CreateDbContext();
 
         var period = Periods.January_EuropeCopenhagen_Instant;
-        var batch = new Batches.Infrastructure.BatchAggregate.Batch(
+        var batch = new Application.BatchAggregate.Batch(
            ProcessType.BalanceFixing,
            new List<GridAreaCode> { new("004") },
            period.PeriodStart,
@@ -227,15 +227,15 @@ public class BatchRepositoryTests : IClassFixture<WholesaleDatabaseFixture>
             actual.Should().NotContain(batch);
     }
 
-    private static Batches.Infrastructure.BatchAggregate.Batch CreateBatch(List<GridAreaCode> someGridAreasIds)
+    private static Application.BatchAggregate.Batch CreateBatch(List<GridAreaCode> someGridAreasIds)
     {
         return CreateBatch(ProcessType.BalanceFixing, someGridAreasIds);
     }
 
-    private static Batches.Infrastructure.BatchAggregate.Batch CreateBatch(ProcessType processType, List<GridAreaCode> someGridAreasIds)
+    private static Application.BatchAggregate.Batch CreateBatch(ProcessType processType, List<GridAreaCode> someGridAreasIds)
     {
         var period = Periods.January_EuropeCopenhagen_Instant;
-        return new Batches.Infrastructure.BatchAggregate.Batch(
+        return new Application.BatchAggregate.Batch(
             processType,
             someGridAreasIds,
             period.PeriodStart,
