+++ source/dotnet/wholesale-api/Batches/Batches.UnitTests/Infrastructure/BatchAggregate/BatchFactoryTests.cs
@@ -14,7 +14,7 @@
 
 using AutoFixture.Xunit2;
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
-using Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
+using Energinet.DataHub.Wholesale.Batches.Application.BatchAggregate;
 using Energinet.DataHub.Wholesale.Batches.Interfaces.Models;
 using FluentAssertions;
 using Moq;
