+++ source/dotnet/wholesale-api/Batches/Batches.UnitTests/Infrastructure/GridAreaAggregate/GridAreaCodeTests.cs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.Wholesale.Batches.Application.GridAreaAggregate;
 using Energinet.DataHub.Wholesale.Batches.Infrastructure;
-using Energinet.DataHub.Wholesale.Batches.Infrastructure.GridAreaAggregate;
 using Energinet.DataHub.Wholesale.Batches.Interfaces;
 using FluentAssertions;
 using Xunit;
