+++ source/dotnet/wholesale-api/WebApi/WebApi.csproj
@@ -36,8 +36,8 @@ limitations under the License.
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Batches\Batches.Infrastructure\Batches.Infrastructure.csproj" />
     <ProjectReference Include="..\Common\Common.csproj" />
-    <ProjectReference Include="..\Batches\Batches.Application\Batches.Application.csproj" />
     <ProjectReference Include="..\CalculationResults\CalculationResults.Application\CalculationResults.Application.csproj" />
     <ProjectReference Include="..\CalculationResults\CalculationResults.Infrastructure\CalculationResults.Infrastructure.csproj" />
     <ProjectReference Include="..\IntegrationEventPublishing\IntegrationEventPublishing.Application\IntegrationEventPublishing.Application.csproj" />
